move all mano related ops from ops -> ops/mano
